+++ vsscale/shaders.py
@@ -8,14 +8,14 @@ from jetpytools import SPath, SPathLike
 from vskernels import Catrom, KernelT, ScalerT
 from vstools import ConstantFormatVideoNode, check_variable, core, depth, inject_self, join, vs
 
-from .helpers import GenericScaler
+from .generic import BaseGenericScaler
 
 __all__ = [
     'PlaceboShader',
 ]
 
 
-class PlaceboShader(GenericScaler):
+class PlaceboShader(BaseGenericScaler):
     """Base placebo shader class."""
 
     _static_kernel_radius = 2
